feat: 컴포넌트 분리 (#81)
- 이메일 전송 컴포넌트를 나누었습니다.
- 지원자의 id를 숫자 배열로 전송해야 하므로 타입을 수정하고 로직을 수정했습니다.
- 지원자 검색 부분은 검색을 한 번 하고 새로 고침을 했을때 화면에 나타납니다. 수정이 필요합니다.
